refactor(site/src/pages/AgentsPage): break AgentChatPage circular dep
AgentChatPageView imported localStorage helpers from AgentChatPage,
which imports AgentChatPageView, closing a cycle that dpdm reports
but does not fail on. Move the helpers into utils/sidebarTabStorage.ts
alongside draftStorage and the other per-chat storage modules.
